move commands daemon into commands folder
