+++ cli/daemon/daemon.go
@@ -25,7 +25,7 @@ import (
 	"runtime"
 
 	"github.com/arduino/arduino-cli/cli"
-	"github.com/arduino/arduino-cli/daemon"
+	"github.com/arduino/arduino-cli/commands/daemon"
 	rpc "github.com/arduino/arduino-cli/rpc/commands"
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
